backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
asyncpg==0.29.0
pydantic==2.8.2
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.2